+++ app/lib/ui/components/components.dart
@@ -4,6 +4,7 @@ import 'package:ods_flutter_demo/ui/components/app_bars/top/top_app_bars.dart';
 import 'package:ods_flutter_demo/ui/components/buttons/button_contained.dart';
 import 'package:ods_flutter_demo/ui/components/buttons/button_outlined.dart';
 import 'package:ods_flutter_demo/ui/components/buttons/button_text.dart';
+import 'package:ods_flutter_demo/ui/components/cards/card_horizontal.dart';
 import 'package:ods_flutter_demo/ui/components/cards/card_small.dart';
 import 'package:ods_flutter_demo/ui/components/cards/card_vertical_image_first.dart';
 import 'package:ods_flutter_demo/ui/components/checkboxes/checkboxes.dart';
@@ -102,6 +103,10 @@ List<Component> components(BuildContext context) {
             AppLocalizations.of(context)!.cardVerticalImageFirstVariantTitle,
             AppLocalizations.of(context)!.cardVerticalImageFirstVariantSubtitle,
             CardVerticalImageFirst()),
+        Variant(
+            AppLocalizations.of(context)!.componentCardHorizontalTitle,
+            AppLocalizations.of(context)!.componentCardHorizontalSubtitle,
+            CardHorizontal()),
       ],
     ),
     Component(