+++ library/lib/components/card/ods_horizontal_card.dart
@@ -1,0 +1,192 @@
+import 'package:flutter/material.dart';
+import 'package:ods_flutter/components/button/ods_button.dart';
+import 'package:ods_flutter/components/button/ods_outlined_button.dart';
+import 'package:ods_flutter/components/card/ods_cards_common.dart';
+import 'package:ods_flutter/guidelines/spacings.dart';
+import 'package:ods_flutter/theme/ods_theme.dart';
+
+enum OdsHorizontalCardImagePosition {
+  start,
+  end,
+}
+
+/// ODS Horizontal card.
+///
+/// Cards contain content and actions about a single subject.
+///
+/// A ripple effect is managed on card click.
+class OdsHorizontalCard extends StatefulWidget {
+  const OdsHorizontalCard({
+    Key? key,
+    required this.title,
+    required this.image,
+    this.subtitle,
+    this.text,
+    this.firstButton,
+    this.secondButton,
+    this.imagePosition = OdsHorizontalCardImagePosition.start,
+    this.divider = true,
+    this.onClick,
+  }) : super(key: key);
+
+  static const double _imageHeight = 0;
+  static const double _imageWidth = 150;
+  static const double _imageMinWidth = 120;
+  static const double _dividerOpacity = 0.12;
+
+  /// The image displayed in the card.
+  ///TODO For the moment the fit of the image is handled by the provided image. It should be done in the library but we need help to do that!
+  final OdsCardImage image;
+
+  /// The card's title displayed below the image.
+  final String title;
+
+  /// The optional card's subtitle displayed below the title.
+  final String? subtitle;
+
+  /// The optional card's text.
+  final String? text;
+
+  /// Optional first button. If null, button will not be shown.
+  final OdsOutlinedButton? firstButton;
+
+  /// Optional second button in the card. If null, button will not be shown.
+  final OdsButton? secondButton;
+
+  /// imagePosition Position of the image within the card, it can be set to [OdsHorizontalCardImagePosition.Start] or [OdsHorizontalCardImagePosition.End]. [OdsHorizontalCardImagePosition.Start] by default.
+  final OdsHorizontalCardImagePosition imagePosition;
+
+  /// Optional divider in the card. If false, divider will not be shown.
+  final bool? divider;
+
+  /// The action executed on card tap.
+  final Function()? onClick;
+
+  @override
+  State<OdsHorizontalCard> createState() => _OdsHorizontalCardState();
+}
+
+class _OdsHorizontalCardState extends State<OdsHorizontalCard> {
+  @override
+  Widget build(BuildContext context) {
+    return Semantics(
+      container: true,
+      image: false,
+      label: 'Card',
+      child: SizedBox(
+        width: double.infinity,
+        child: Material(
+          color: Colors.transparent,
+          child: InkWell(
+            onTap: widget.onClick,
+            child: Card(
+              shape: RoundedRectangleBorder(
+                borderRadius: BorderRadius.circular(odsCardRadius),
+              ),
+              elevation: 0.3,
+              clipBehavior: Clip.antiAliasWithSaveLayer,
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  Container(
+                    constraints: const BoxConstraints(
+                        minHeight: OdsHorizontalCard._imageMinWidth),
+                    child: IntrinsicHeight(
+                      child: Row(
+                        crossAxisAlignment: CrossAxisAlignment.stretch,
+                        children: [
+                          if (widget.imagePosition ==
+                              OdsHorizontalCardImagePosition.start)
+                            ExcludeSemantics(
+                              child: SizedBox(
+                                width: OdsHorizontalCard._imageWidth,
+                                height: OdsHorizontalCard._imageHeight,
+                                child: widget.image,
+                              ),
+                            ),
+                          Expanded(
+                            child: Padding(
+                              padding: const EdgeInsets.only(
+                                  left: spacingM,
+                                  top: spacingM,
+                                  right: spacingM,
+                                  bottom: spacingS),
+                              child: Column(
+                                crossAxisAlignment: CrossAxisAlignment.start,
+                                children: [
+                                  Text(
+                                    widget.title,
+                                    style:
+                                        Theme.of(context).textTheme.titleLarge,
+                                  ),
+                                  if (widget.subtitle != null)
+                                    Text(
+                                      widget.subtitle!,
+                                      style: Theme.of(context)
+                                          .textTheme
+                                          .bodyMedium,
+                                    ),
+                                  if (widget.text != null)
+                                    Padding(
+                                      padding:
+                                          const EdgeInsets.only(top: spacingM),
+                                      child: Text(widget.text!,
+                                          style: Theme.of(context)
+                                              .textTheme
+                                              .bodyMedium,
+                                          maxLines: 2,
+                                          overflow: TextOverflow.ellipsis),
+                                    ),
+                                ],
+                              ),
+                            ),
+                          ),
+                          if (widget.imagePosition ==
+                              OdsHorizontalCardImagePosition.end)
+                            ExcludeSemantics(
+                              child: SizedBox(
+                                width: OdsHorizontalCard._imageWidth,
+                                height: OdsHorizontalCard._imageHeight,
+                                child: widget.image,
+                              ),
+                            )
+                        ],
+                      ),
+                    ),
+                  ),
+                  if (widget.divider != false)
+                    Divider(
+                      height: 0,
+                      thickness: 1,
+                      indent: 0,
+                      color: Theme.of(context)
+                          .colorScheme
+                          .onSurface
+                          .withOpacity(OdsHorizontalCard._dividerOpacity),
+                    ),
+                  SingleChildScrollView(
+                    scrollDirection: Axis.horizontal,
+                    child: Row(
+                      children: [
+                        if (widget.firstButton != null)
+                          Padding(
+                            padding: const EdgeInsets.all(spacingS),
+                            child: widget.firstButton,
+                          ),
+                        if (widget.secondButton != null)
+                          Padding(
+                            padding: const EdgeInsets.all(spacingS),
+                            child: widget.secondButton,
+                          ),
+                      ],
+                    ),
+                  ),
+                ],
+              ),
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+}